+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+
+- Changed the `name` variable / attribute to `prefix` to follow uniformity across   all executor plugins
+- Changed the **dict** type of defaults to `ExecutorInfraDefaults` & `ExecutorPluginDefaults`
+- Minor changes in `MANIFEST.in` file
+
+### Added
+
+- Added a tftpl file for ECS executor to output the configuration
+
 ## [0.32.0] - 2023-10-06
 
 ### Changed